requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
Django==4.0.6
django-crispy-forms==1.14.0
django-gravatar2==1.4.4
django-hashid-field==3.3.5
django-phonenumber-field==6.3.0
django-tinymce==3.4.0
django-widget-tweaks==1.4.12
hashids==1.3.1
idna==3.3
isodate==0.6.1
lxml==4.9.1
phonenumbers==8.12.51
phonenumberslite==8.12.51
Pillow==9.2.0
platformdirs==2.5.2
pycryptodome==3.15.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
suds-py3==1.4.5.0
urllib3==1.26.11
zeep==3.4.0
